create turso_parser crate
