+++ core/Cargo.toml
@@ -111,7 +111,3 @@ harness = false
 [[bench]]
 name = "tpc_h_benchmark"
 harness = false
-
-[[bench]]
-name = "parser_benchmark"
-harness = false
