+++ parser/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "turso_parser"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "The Turso parser library"
+
+[lib]
+name = "turso_parser"
+
+[features]
+default = []
+serde = ["dep:serde", "bitflags/serde"]
+
+[dependencies]
+bitflags = "2.0"
+miette = "7.4.0"
+strum = { workspace = true }
+strum_macros = {workspace = true }
+serde = { workspace = true , optional = true, features = ["derive"] }
+
+[dev-dependencies]
+fallible-iterator = "0.3"
+criterion = { version = "0.5", features = ["html_reports" ] }
+turso_sqlite3_parser = { workspace = true }
+
+[target.'cfg(not(target_family = "windows"))'.dev-dependencies]
+pprof = { version = "0.14.0", features = ["criterion", "flamegraph"] }
+
+[[bench]]
+name = "parser_benchmark"
+harness = false
